FIXED: bad commit, return to revision 20
git-svn-id: svn://svn.code.sf.net/p/nel/code/trunk@22 bf685bcc-7839-4fcd-9254-36823579bede

+++ nel/src/3d/driver/opengl/driver_opengl_vertex_program.cpp
@@ -575,7 +575,7 @@ bool CDriverGL::setupEXTVertexShader(const CVPParser::TProgram &program, GLuint 
 					}
 					break;
 					case CVPOperand::Constant: 
-						nlassert(uint(operand.Value.ConstantValue) < _EVSNumConstant); // constant index too high
+						nlassert(operand.Value.ConstantValue < _EVSNumConstant); // constant index too high
 						srcValue[l] = _EVSConstantHandle + operand.Value.ConstantValue; 												
 						EVS_INFO(("Src = constant" + toString(operand.Value.ConstantValue)).c_str());
 					break;
